# internal/openapi/swagger.yaml
basePath: /api
...
info:
  contact: {}
  description: HTTP API for the AWG Manager daemon (tunnels, routing, system).
  title: AWG Manager API
  version: "1.0"
paths:
...
  /control/restart:
    post:
      parameters:
      - description: Tunnel id
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - CookieAuth: []
      summary: Restart tunnel
      tags:
      - control
  /control/restart-all:
    post:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
      security:
      - CookieAuth: []
      summary: Restart all enabled tunnels
      tags:
      - control
  /control/start:
    post:
      parameters:
      - description: Tunnel id
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - CookieAuth: []
      summary: Start tunnel
      tags:
      - control
  /control/stop:
    post:
      parameters:
      - description: Tunnel id
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - CookieAuth: []
      summary: Stop tunnel
      tags:
      - control
  /control/toggle-default-route:
    post:
      parameters:
      - description: Tunnel id
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - CookieAuth: []
      summary: Toggle default route
      tags:
      - control
  /control/toggle-enabled:
    post:
      parameters:
      - description: Tunnel id
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - CookieAuth: []
      summary: Toggle tunnel autostart
      tags:
      - control
  /diagnostics/result:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: file
      security:
      - CookieAuth: []
      summary: Download diagnostics report
      tags:
      - diagnostics
  /diagnostics/run:
    post:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - CookieAuth: []
      summary: Run diagnostics
      tags:
      - diagnostics
  /diagnostics/status:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - CookieAuth: []
      summary: Diagnostics status
      tags:
      - diagnostics
  /diagnostics/stream:
    get:
      produces:
      - text/event-stream
      responses:
        "200":
          description: SSE stream
          schema:
            type: string
      security:
      - CookieAuth: []
      summary: Diagnostics SSE stream
      tags:
      - diagnostics
...
  /dns-check/start:
    post:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - CookieAuth: []
      summary: Start DNS check
      tags:
      - dns-check
...
securityDefinitions:
  CookieAuth:
    description: Session cookie set by POST /auth/login. Omit for public routes.
    in: cookie
    name: awg_session
    type: apiKey
swagger: "2.0"